+++ packages/web-api-client/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "web-api-client-library",
+    "name": "web-api-client",
     "version": "1.0.0",
     "description": "This project welcomes contributions and suggestions.  Most contributions require you to agree to a Contributor License Agreement (CLA) declaring that you have the right to, and actually do, grant us the rights to use your contribution. For details, visit https://cla.microsoft.com.",
     "main": "dist/index.js",
