+++ packages/web-api-client/src/index.ts
@@ -1,3 +1,3 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-export { registerWebApiClientLibraryToContainer } from './register-web-api-client-library-to-container';
+export { registerWebApiClientLibraryToContainer } from './register-web-api-client-to-container';
